web/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==3.1.1
djangorestframework==3.11.1
django-filter==2.3.0
django-tinymce==2.8.0
idna==2.8
numpy==1.16.4
pandas==0.25.0
django-pandas
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-cache==0.5.0
psycopg2-binary==2.8.3
six==1.12.0
sqlparse==0.3.0
transliterate==1.10.2
urllib3==1.25.3
supervisor==4.2.1
handy==0.7
gunicorn==20.0.4
django-silk
redis==3.5.3
django-redis==4.12.1
celery[redis]==4.4.7
django-debug-toolbar
django-celery-beat
scipy==1.2.2
django-db-logger==0.1.10
eventlet==0.26.1
dnspython==1.16.0
aiohttp==3.6.2
asyncio==3.4.3
async-timeout==3.0.1
sentry-sdk==0.17.6
django-extensions==3.0.8
ipython==7.18.1
python-json-logger==0.1.11
pyTelegramBotAPI==3.7.3
python-logstash==0.4.6
